fixes #13: separa os testes em múltiplos arquivos
