+++ .github/workflows/docker-image.yml
@@ -21,4 +21,4 @@ jobs:
     - name: Run docker-compose
       run: docker-compose up -d
     - name: run tests
-      run: docker exec api-pgd_web_1 /bin/bash -c "pytest test_api.py"
+      run: docker exec api-pgd_web_1 /bin/bash -c "pytest tests/"
