+++ tests/plano_trabalho_test.py
@@ -1,146 +1,13 @@
 """
-Testes automáticos da API
+Testes relacionados ao plano de trabalho.
 """
-import os, subprocess
-import json
 import itertools
-from typing import Callable, Generator, Optional
-from datetime import datetime
 
-from requests.models import Response as HTTPResponse
-import requests
-
-from fastapi.testclient import TestClient
 from requests import Session
-from fastapi import status
-from api import app
 
-import util
+from fastapi import status
 
 import pytest
-
-# Helper functions
-
-def register_user(
-        client: Session,
-        email: str,
-        password: str,
-        cod_unidade: int,
-        headers: dict
-    ) -> HTTPResponse:
-    data = {
-        "email": email,
-        "password": password,
-        "cod_unidade": cod_unidade,
-    }
-    return client.post(
-        f"/auth/register",
-        json=data,
-        headers=headers
-    )
-
-def prepare_header(username: Optional[str], password: Optional[str]) -> dict:
-    """Prepara o cabeçalho para ser utilizado em requisições.
-    """
-    #TODO: Refatorar e resolver utilizando o objeto TestClient
-    token_user = None
-
-    if username and password:
-        # usuário especificado, é necessário fazer login
-        url = "http://localhost:5057/auth/jwt/login"
-
-        headers = {
-        "Content-Type": "application/x-www-form-urlencoded"
-        }
-
-        payload="&".join([
-            "accept=application%2Fjson",
-            "Content-Type=application%2Fjson",
-            f"username={username}",
-            f"password={password}"
-        ])
-
-        response = requests.request("POST", url, headers=headers, data=payload)
-        response_dict = json.loads(response.text)
-        token_user = response_dict.get("access_token")
-        print(token_user)
-
-    headers = {
-        "accept": "application/json",
-        "Content-Type": "application/json"
-    }
-    if token_user:
-        headers["Authorization"] = f"Bearer {token_user}"
-
-    return headers
-
-
-# Fixtures
-
-@pytest.fixture(scope="module")
-def client() -> Generator[Session, None, None]:
-    with TestClient(app) as c:
-        yield c
-
-@pytest.fixture()
-def input_pt() -> dict:
-    pt_json = {
-        "cod_plano": "555",
-        "matricula_siape": 0,
-        "cpf": "99160773120",
-        "nome_participante": "string",
-        "cod_unidade_exercicio": 0,
-        "nome_unidade_exercicio": "string",
-        "modalidade_execucao": 1,
-        "carga_horaria_semanal": 10,
-        "data_inicio": "2021-01-07",
-        "data_fim": "2021-01-12",
-        "carga_horaria_total": 0.0,
-        "data_interrupcao": "2021-01-07",
-        "entregue_no_prazo": True,
-        "horas_homologadas": 0.0,
-        "atividades": [
-            {
-                "id_atividade": "2",
-                "nome_grupo_atividade": "string",
-                "nome_atividade": "string",
-                "faixa_complexidade": "string",
-                "parametros_complexidade": "string",
-                "tempo_presencial_estimado": 0.0,
-                "tempo_presencial_programado": 0.0,
-                "tempo_presencial_executado": None,
-                "tempo_teletrabalho_estimado": 0.0,
-                "tempo_teletrabalho_programado": 0.0,
-                "tempo_teletrabalho_executado": None,
-                "entrega_esperada": "string",
-                "qtde_entregas": 0,
-                "qtde_entregas_efetivas": 0,
-                "avaliacao": 0,
-                "data_avaliacao": "2021-01-15",
-                "justificativa": "string"
-            },
-            {
-                "id_atividade": "3",
-                "nome_grupo_atividade": "string",
-                "nome_atividade": "string",
-                "faixa_complexidade": "string",
-                "parametros_complexidade": "string",
-                "tempo_presencial_estimado": 0.0,
-                "tempo_presencial_programado": 0.0,
-                "tempo_presencial_executado": None,
-                "tempo_teletrabalho_estimado": 0.0,
-                "tempo_teletrabalho_programado": 0.0,
-                "tempo_teletrabalho_executado": None,
-                "entrega_esperada": "string",
-                "qtde_entregas": 0,
-                "qtde_entregas_efetivas": 0,
-                "avaliacao": 0,
-                "data_avaliacao": "2021-01-15",
-                "justificativa": "string"
-            }
-        ]
-    }
-    return pt_json
 
 # grupos de campos opcionais e obrigatórios a testar
 
@@ -164,233 +31,6 @@ fields_plano_trabalho = {
         ["atividades"],
     )
 }
-
-fields_atividade = {
-    "optional": (
-        ["nome_grupo_atividade"],
-        ["parametros_complexidade"],
-        ["entrega_esperada"],
-        ["qtde_entregas_efetivas"],
-        ["data_avaliacao"],
-        ["justificativa"],
-        ["tempo_presencial_executado"],
-        ["tempo_teletrabalho_executado"],
-    ),
-    "mandatory": (
-        ["nome_atividade"],
-        ["faixa_complexidade"],
-        ["tempo_presencial_estimado"],
-        ["tempo_presencial_programado"],
-        ["tempo_teletrabalho_estimado"],
-        ["tempo_teletrabalho_programado"],
-        ["qtde_entregas"],
-    )
-}
-
-atividades_dict = {
-    "2": {
-        "nome_grupo_atividade": "string",
-        "nome_atividade": "string",
-        "faixa_complexidade": "string",
-        "parametros_complexidade": "string",
-        "tempo_presencial_estimado": 0.0,
-        "tempo_presencial_programado": 0.0,
-        "tempo_presencial_executado": None,
-        "tempo_teletrabalho_estimado": 0.0,
-        "tempo_teletrabalho_programado": 0.0,
-        "tempo_teletrabalho_executado": None,
-        "entrega_esperada": "string",
-        "qtde_entregas": 0,
-        "qtde_entregas_efetivas": 0,
-        "avaliacao": 0,
-        "data_avaliacao": "2021-01-15",
-    "justificativa": "string"
-    },
-    "3": {
-        "nome_grupo_atividade": "string",
-        "nome_atividade": "string",
-        "faixa_complexidade": "string",
-        "parametros_complexidade": "string",
-        "tempo_presencial_estimado": 0.0,
-        "tempo_presencial_programado": 0.0,
-        "tempo_presencial_executado": None,
-        "tempo_teletrabalho_estimado": 0.0,
-        "tempo_teletrabalho_programado": 0.0,
-        "tempo_teletrabalho_executado": None,
-        "entrega_esperada": "string",
-        "qtde_entregas": 0,
-        "qtde_entregas_efetivas": 0,
-        "avaliacao": 0,
-        "data_avaliacao": "2021-01-15",
-        "justificativa": "string"
-    }
-}
-
-@pytest.fixture(scope="module")
-def admin_credentials() -> dict:
-    return {
-        "username": "admin@api.com",
-        "password": "1234",
-        "cod_unidade": 1
-    }
-
-@pytest.fixture(scope="module")
-def user1_credentials() -> dict:
-    return {
-        "username": "test1@api.com",
-        "password": "api",
-        "cod_unidade": 1
-    }
-
-@pytest.fixture(scope="module")
-def user2_credentials() -> dict:
-    return {
-        "username": "test2@api.com",
-        "password": "api",
-        "cod_unidade": 2
-    }
-
-@pytest.fixture()
-def example_pt(client: Session, input_pt: dict, header_usr_1: dict):
-    client.put(f"/plano_trabalho/555",
-                          json=input_pt,
-                          headers=header_usr_1)
-
-@pytest.fixture()
-def truncate_pt(client: Session, header_admin: dict):
-    client.post(f"/truncate_pts_atividades", headers=header_admin)
-
-@pytest.fixture(scope="module")
-def truncate_users():
-    p = subprocess.Popen(
-        [
-            "/usr/local/bin/python",
-            "/home/api-pgd/admin_tool.py",
-            "--truncate-users"
-        ],
-        stdout=subprocess.PIPE,
-        stdin=subprocess.PIPE,
-        stderr=subprocess.PIPE,
-        text=True
-    )
-
-@pytest.fixture(scope="module")
-def register_admin(truncate_users, admin_credentials: dict):
-    email = admin_credentials["username"]
-    cod_unidade = admin_credentials["cod_unidade"]
-    password = admin_credentials["password"]
-    p = subprocess.Popen(
-        [
-            "/usr/local/bin/python",
-            "/home/api-pgd/admin_tool.py",
-            "--create_superuser",
-            "--show_password"
-        ],
-        stdout=subprocess.PIPE,
-        stdin=subprocess.PIPE,
-        stderr=subprocess.PIPE,
-        text=True
-    )
-    p.communicate(input="\n".join([
-        email, str(cod_unidade), password, password
-    ]))[0]
-
-@pytest.fixture(scope="module")
-def register_user_1(
-    client: Session,
-    truncate_users,
-    register_admin,
-    header_admin: dict,
-    user1_credentials: dict
-    ) -> HTTPResponse:
-    return register_user(client, user1_credentials["username"],
-        user1_credentials["password"], user1_credentials["cod_unidade"],
-        header_admin)
-
-@pytest.fixture(scope="module")
-def register_user_2(
-    client: Session,
-    truncate_users,
-    register_admin,
-    header_admin: dict,
-    user2_credentials: dict
-    ) -> HTTPResponse:
-    return register_user(client, user2_credentials["username"],
-        user2_credentials["password"], user2_credentials["cod_unidade"],
-        header_admin)
-
-@pytest.fixture(scope="module")
-def header_not_logged_in() -> dict:
-    return prepare_header(username=None, password=None)
-
-@pytest.fixture(scope="module")
-def header_admin(register_admin, admin_credentials: dict) -> dict:
-    return prepare_header(
-        username=admin_credentials["username"],
-        password=admin_credentials["password"])
-
-@pytest.fixture(scope="module")
-def header_usr_1(register_user_1, user1_credentials: dict) -> dict:
-    """Authenticate in the API as user1 and return a dict with bearer
-    header parameter to be passed to apis requests."""
-    return prepare_header(
-        username=user1_credentials["username"],
-        password=user1_credentials["password"])
-
-@pytest.fixture(scope="module")
-def header_usr_2(register_user_2, user2_credentials: dict) -> dict:
-    """Authenticate in the API as user2 and return a dict with bearer
-    header parameter to be passed to apis requests."""
-    return prepare_header(
-        username=user2_credentials["username"],
-        password=user2_credentials["password"])
-
-
-# Tests
-
-# API Tests
-
-def test_register_user_not_logged_in(
-        truncate_users, client: Session, header_not_logged_in: dict):
-    user_1 = register_user(client, "testx@api.com", "api", 0, header_not_logged_in)
-    assert user_1.status_code == status.HTTP_401_UNAUTHORIZED
-
-def test_register_user(truncate_users, client: Session, header_admin: dict):
-    user_1 = register_user(client, "testx@api.com", "api", 0, header_admin)
-    assert user_1.status_code == status.HTTP_201_CREATED
-
-    user_2 = register_user(client, "testx@api.com", "api", 0, header_admin)
-    assert user_2.status_code == status.HTTP_400_BAD_REQUEST
-    assert user_2.json().get("detail", None) == "REGISTER_USER_ALREADY_EXISTS"
-
-def test_authenticate(header_usr_1: dict):
-    token = header_usr_1.get("Authorization")
-    assert isinstance(token, str)
-    assert "eyJ0eXAiOiJKV1QiLCJhbGciOiJIUzI1NiJ9." in token
-
-def test_get_user_self_not_logged_in(client: Session,
-        header_not_logged_in: dict):
-    response = client.get("/users/me", headers=header_not_logged_in)
-    assert response.status_code == status.HTTP_401_UNAUTHORIZED
-
-def test_get_user_self_logged_in(client: Session, user1_credentials: dict,
-        header_usr_1: dict):
-    response = client.get("/users/me", headers=header_usr_1)
-    assert response.status_code == status.HTTP_200_OK
-    data = response.json()
-    assert data.get("email", None) == user1_credentials["username"]
-    assert data.get("cod_unidade", None) == user1_credentials["cod_unidade"]
-    assert data.get("is_active", None) == True
-
-def test_patch_user_self_change_cod_unidade(client: Session,
-        header_usr_1: dict):
-    " Testa se o usuário pode alterar o seu próprio cod_unidade."
-    response = client.patch(
-        "/users/me",
-        json={"cod_unidade": 3},
-        headers=header_usr_1
-    )
-    assert response.status_code == status.HTTP_401_UNAUTHORIZED
 
 def test_create_plano_trabalho_completo(input_pt: dict,
                                         header_usr_1: dict,
@@ -500,187 +140,6 @@ def test_patch_plano_trabalho_inexistente(truncate_pt,
                           headers=header_usr_1)
     assert response.status_code == status.HTTP_404_NOT_FOUND
 
-@pytest.mark.parametrize("omitted_fields",
-                         enumerate(fields_atividade["optional"]))
-def test_create_atvidades_omit_optional_fields(input_pt: dict,
-                                             omitted_fields: list,
-                                             header_usr_1: dict,
-                                             truncate_pt,
-                                             client: Session):
-    """Tenta criar um plano de trabalho, mas em cada atividade os
-    campos opcionais são omitidos.
-    """
-    offset, field_list = omitted_fields
-    for field in field_list:
-        for atividade in input_pt["atividades"]:
-            del atividade[field]
-
-    input_pt["cod_plano"] = 557 + offset
-    response = client.put(f"/plano_trabalho/{input_pt['cod_plano']}",
-                          json=input_pt,
-                          headers=header_usr_1)
-    print(response.json())
-    assert response.status_code == status.HTTP_200_OK
-
-@pytest.mark.parametrize("missing_fields",
-                         enumerate(fields_atividade["mandatory"]))
-def test_create_atividades_missing_mandatory_fields(input_pt: dict,
-                                             missing_fields: list,
-                                             header_usr_1: dict,
-                                             truncate_pt,
-                                             client: Session):
-    """Tenta criar um plano de trabalho, mas em cada atividade os
-    campos obrigatórios são omitidos.
-    """
-    offset, field_list = missing_fields
-    for field in field_list:
-        for atividade in input_pt["atividades"]:
-            del atividade[field]
-
-    input_pt["cod_plano"] = 1800 + offset # precisa ser um novo plano
-    response = client.put(f"/plano_trabalho/{input_pt['cod_plano']}",
-                          json=input_pt,
-                          headers=header_usr_1)
-    assert response.status_code == status.HTTP_422_UNPROCESSABLE_ENTITY
-
-@pytest.mark.parametrize("verb, missing_fields",
-                    itertools.product( # todas as combinações entre
-                        ("put", "patch"), # verb
-                        fields_atividade["mandatory"] # missing_fields
-                    ))
-def test_update_atividades_missing_mandatory_fields(verb: str,
-                                            truncate_pt,
-                                            example_pt,
-                                            input_pt: dict,
-                                            missing_fields: list,
-                                            header_usr_1: dict,
-                                            client: Session):
-    """Tenta atualizar as atividades de um plano de trabalho faltando
-    campos obrigatórios.
-
-    Para usar o verbo PUT é necessário passar a representação completa
-    do recurso. Por isso, os campos obrigatórios não podem estar
-    faltando. Para realizar uma atualização parcial, use o método PATCH.
-
-    Já com o verbo PATCH, os campos omitidos são interpretados como sem
-    alteração. Por isso, é permitido omitir os campos obrigatórios.
-    """
-    for field in missing_fields:
-        for atividade in input_pt["atividades"]:
-            del atividade[field]
-
-    input_pt["cod_plano"] = 555 # precisa ser um plano existente
-    call = getattr(client, verb) # put ou patch
-    response = call(f"/plano_trabalho/{input_pt['cod_plano']}",
-                          json=input_pt,
-                          headers=header_usr_1)
-    if verb == "put":
-        assert response.status_code == status.HTTP_422_UNPROCESSABLE_ENTITY
-    elif verb == "patch":
-        assert response.status_code == status.HTTP_200_OK
-
-def test_substitute_atividades_list(truncate_pt,
-                                    example_pt,
-                                    input_pt: dict,
-                                    header_usr_1: dict,
-                                    client: Session):
-    "Substitui a lista de atividades existentes por uma nova lista."
-    input_pt["atividades"].pop() # remove a última atividade
-
-    response = client.put(f"/plano_trabalho/{input_pt['cod_plano']}",
-                            json=input_pt,
-                            headers=header_usr_1)
-
-    assert response.status_code == status.HTTP_200_OK
-    assert response.json() == input_pt
-
-def test_append_atividades_list(truncate_pt,
-                                example_pt,
-                                input_pt: dict,
-                                header_usr_1: dict,
-                                client: Session):
-    "Acrescenta uma nova atividade à lista de atividades existentes."
-    nova_atividade = {
-      "id_atividade": "4",
-      "nome_grupo_atividade": "string",
-      "nome_atividade": "string",
-      "faixa_complexidade": "string",
-      "parametros_complexidade": "string",
-      "tempo_presencial_estimado": 0.0,
-      "tempo_presencial_programado": 0.0,
-      "tempo_presencial_executado": None,
-      "tempo_teletrabalho_estimado": 0.0,
-      "tempo_teletrabalho_programado": 0.0,
-      "tempo_teletrabalho_executado": None,
-      "entrega_esperada": "string",
-      "qtde_entregas": 0,
-      "qtde_entregas_efetivas": 0,
-      "avaliacao": 0,
-      "data_avaliacao": "2021-02-15",
-      "justificativa": "string"
-    }
-
-    patch_input = {
-        "cod_plano": input_pt["cod_plano"],
-        "atividades": [nova_atividade]
-    }
-
-    response = client.patch(f"/plano_trabalho/{input_pt['cod_plano']}",
-                            json=patch_input,
-                            headers=header_usr_1)
-
-    input_pt["atividades"].append(nova_atividade)
-
-    assert response.status_code == status.HTTP_200_OK
-    assert response.json() == input_pt
-
-@pytest.mark.parametrize("verb, data_avaliacao",
-                            itertools.product(
-                                ("put", "patch"),
-                                (
-                                    "2021-01-11", # < data_fim do plano
-                                    "2021-03-01"
-                                )
-                            )
-                        )
-def test_modify_atividade(truncate_pt,
-                            example_pt,
-                            verb: str,
-                            data_avaliacao: str,
-                            input_pt: dict,
-                            header_usr_1: dict,
-                            client: Session):
-    "Modifica uma atividade existente com put e patch."
-    atividade = input_pt["atividades"][-1] # pega a última atividade
-    atividade["data_avaliacao"] = data_avaliacao
-
-    if verb == "put":
-        response = client.put(f"/plano_trabalho/{input_pt['cod_plano']}",
-                          json=input_pt,
-                          headers=header_usr_1)
-    elif verb == "patch":
-        atividade_patch = {
-            "cod_plano": input_pt["cod_plano"],
-            "atividades": [
-                {
-                    "id_atividade": "3",
-                    "data_avaliacao": data_avaliacao
-                }
-            ]
-        }
-        response = client.patch(f"/plano_trabalho/{input_pt['cod_plano']}",
-                          json=atividade_patch,
-                          headers=header_usr_1)
-
-    if datetime.fromisoformat(data_avaliacao) < datetime.fromisoformat(
-                                                    input_pt["data_fim"]):
-        assert response.status_code == status.HTTP_422_UNPROCESSABLE_ENTITY
-        detail_msg = "Data de avaliação da atividade deve ser maior ou igual" \
-                     " que a Data Fim do Plano de Trabalho."
-        assert response.json().get("detail")[0]["msg"] == detail_msg
-    else:
-        assert response.status_code == status.HTTP_200_OK
-        assert response.json() == input_pt
 
 def test_create_pt_cod_plano_inconsistent(input_pt: dict,
                                           header_usr_1: dict,
@@ -984,30 +443,3 @@ def test_create_pt_missing_mandatory_fields_atividade(input_pt: dict,
     assert response.status_code == status.HTTP_422_UNPROCESSABLE_ENTITY
     detail_msg = "none is not an allowed value"
     assert response.json().get("detail")[0]["msg"] == detail_msg
-
-# Unit tests
-
-def test_merge_dicts(input_pt: dict):
-    """Testa a mesclagem de dicionários.
-    """
-    input1 = input_pt.copy()
-    input2 = input_pt.copy()
-
-    del input1["atividades"][0]["qtde_entregas"]
-    del input2["atividades"][0]["avaliacao"]
-
-    assert util.merge_dicts(input1, input2) == input_pt
-
-def test_list_to_dict(input_pt: dict):
-    """Testa a transformação de lista em dicionário.
-    """
-    atividades = util.list_to_dict(input_pt["atividades"], "id_atividade")
-
-    assert atividades == atividades_dict
-
-def test_dict_to_list(input_pt: dict):
-    """Testa a transformação de lista em dicionário.
-    """
-    atividades = util.dict_to_list(atividades_dict, "id_atividade")
-
-    assert atividades == input_pt["atividades"]
